Django/requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
beautifulsoup4==4.11.1
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
channels==2.4.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
constantly==15.1.0
contourpy==1.0.5
cryptography==38.0.1
cvxpy==1.2.1
cycler==0.11.0
daphne==2.5.0
dash==2.6.2
dash-bootstrap-components==0.13.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
DateTime==4.7
Deprecated==1.2.13
dj-database-url==1.0.0
Django==3.2.16
django-plotly-dash==2.0.0
dpd-components==0.1.0
ecos==2.0.10
exceptiongroup==1.0.0rc9
Flask==2.1.3
Flask-Compress==1.13
fonttools==4.37.4
gunicorn==20.1.0
h11==0.14.0
handler==1.3.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.0.0
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.1
multitasking==0.0.11
numpy==1.23.4
osqp==0.6.2.post5
outcome==1.2.0
packaging==21.3
pandas==1.5.0
pandas-datareader==0.10.0
Pillow==9.2.0
plotly==5.10.0
psycopg2==2.9.4
psycopg2-binary==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pykrx==1.0.38
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyperclip==1.8.2
pyportfolioopt==1.5.3
pyproj==3.4.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.4
PyYAML==6.0
qdldl==0.1.5.post2
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.3
scs==3.2.0
selenium==4.5.0
service-identity==21.1.0
six==1.16.0
sklearn==0.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tenacity==8.1.0
threadpoolctl==3.1.0
trio==0.22.0
trio-websocket==0.9.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.4
urllib3==1.26.12
Werkzeug==2.0.3
whitenoise==6.2.0
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1
yfinance==0.1.81
zipp==3.9.0
zope.interface==5.5.0
mysqlclient
pyOpenSSL
oracledb
django-environ
django-cors-headers